refactor: extract ensureDynamicLanguages to ast-languages.ts
Remove duplicate dynamic language registration from chunker.ts and
graph-imports.ts, consolidating into a single shared module.
